+++ content/ui/__tests__/LeetSrsControl.test.tsx
@@ -3,7 +3,7 @@ import { act, fireEvent, render, screen, waitFor } from '@testing-library/react'
 import { expect, it, vi } from 'vitest';
 import { translations } from '@/i18n';
 import { createDeferred } from '@/test/utils/deferred';
-import { LeetSrsControl } from '../LeetSrsControl';
+import { LeetSrsButton, LeetSrsControl } from '../LeetSrsControl';
 
 function setup(getTranslations = vi.fn().mockResolvedValue(translations.en)) {
   const onRate = vi.fn();
@@ -65,4 +65,19 @@ it('owns the tooltip portal and removes it on mouse leave or unmount', async () 
   await screen.findByRole('tooltip');
   unmount();
   await waitFor(() => expect(screen.queryByRole('tooltip')).not.toBeInTheDocument());
+});
+
+it('renders an accessible toolbar button and dispatches once per click', () => {
+  const onClick = vi.fn();
+  const { rerender } = render(<LeetSrsButton t={translations.en} onClick={onClick} />);
+  const button = screen.getByRole('button', { name: translations.en.app.name });
+  expect(button).toHaveAttribute('type', 'button');
+  expect(button).toHaveStyle({ color: '#28c244' });
+  fireEvent.click(button);
+  expect(onClick).toHaveBeenCalledOnce();
+  const nextClick = vi.fn();
+  rerender(<LeetSrsButton t={translations.pl} onClick={nextClick} />);
+  fireEvent.click(screen.getByRole('button', { name: translations.pl.app.name }));
+  expect(nextClick).toHaveBeenCalledOnce();
+  expect(onClick).toHaveBeenCalledOnce();
 });
