+++ content/ui/__tests__/theme.test.tsx
@@ -1,11 +1,11 @@
 // @vitest-environment happy-dom
 import { act, render, screen, waitFor } from '@testing-library/react';
-import { afterEach, expect, it, vi } from 'vitest';
+import { afterEach, describe, expect, it, vi } from 'vitest';
 import { translations } from '@/i18n';
-import { THEME_COLORS } from '../constants';
-import { LeetSrsButton } from '../LeetSrsButton';
+import { LeetSrsButton } from '../LeetSrsControl';
 import { RatingMenu } from '../RatingMenu';
 import { Tooltip } from '../Tooltip';
+import { getRatingColor, isDarkMode, type RATING_COLORS, THEME_COLORS } from '../theme';
 
 afterEach(() => {
   document.documentElement.className = '';
@@ -50,4 +50,35 @@ it('disconnects theme observers on unmount', () => {
   const { unmount } = render(<LeetSrsButton t={translations.en} onClick={vi.fn()} />);
   unmount();
   expect(disconnect).toHaveBeenCalledOnce();
+});
+
+describe('theme utilities', () => {
+  describe('isDarkMode', () => {
+    it.each([
+      [false, false],
+      [true, true],
+    ])('should return %s when dark mode is %s', (expected, dark) => {
+      document.documentElement.classList.toggle('dark', dark);
+      expect(isDarkMode()).toBe(expected);
+    });
+  });
+
+  describe('getRatingColor', () => {
+    it.each([
+      ['again', false, '#c73e3e', '#b13636'],
+      ['again', true, '#d14358', '#c13a4f'],
+      ['hard', false, '#d97706', '#c26805'],
+      ['good', false, '#4271c4', '#3862b5'],
+      ['easy', false, '#3d9156', '#35804a'],
+    ] as const)('should return the %s colors when dark mode is %s', (rating, dark, bg, hover) => {
+      document.documentElement.classList.toggle('dark', dark);
+      expect(getRatingColor(rating)).toEqual({ bg, hover });
+    });
+
+    it('should throw error for unknown color class', () => {
+      // Since getRatingColor expects a valid key of RATING_COLORS,
+      // it would throw an error for unknown keys
+      expect(() => getRatingColor('unknown-class' as keyof typeof RATING_COLORS)).toThrow();
+    });
+  });
 });
