+++ content/ui/ui.module.css
@@ -73,11 +73,13 @@ button.add:hover {
   box-shadow: 0 4px 6px rgb(0 0 0 / 10%);
   opacity: 0;
   transform: translateX(-50%) translateY(-4px) scale(0.95);
-  transition: all 100ms cubic-bezier(0.16, 1, 0.3, 1);
+  animation: tooltip-enter 100ms cubic-bezier(0.16, 1, 0.3, 1) forwards;
   z-index: 50;
 }
 
-.tooltip[data-visible="true"] {
-  opacity: 1;
-  transform: translateX(-50%) translateY(0) scale(1);
+@keyframes tooltip-enter {
+  to {
+    opacity: 1;
+    transform: translateX(-50%) translateY(0) scale(1);
+  }
 }
